+++ lib/i18n/tasks/command/commands/missing.rb
@@ -36,11 +36,16 @@ module I18n::Tasks
         cmd :translate_missing,
             pos:  '[locale ...]',
             desc: t('i18n_tasks.cmd.desc.translate_missing'),
-            args: [:locales, :locale_to_translate_from, arg(:out_format).from(1)]
+            args: [:locales, :locale_to_translate_from, arg(:out_format).from(1), :translation_backend]
 
         def translate_missing(opt = {})
           missing    = i18n.missing_diff_forest opt[:locales], opt[:from]
-          translated = i18n.google_translate_forest missing, opt[:from]
+          translated = case opt[:backend]
+                       when 'deepl'
+                         i18n.deepl_translate_forest missing, opt[:from]
+                       when 'google'
+                         i18n.google_translate_forest missing, opt[:from]
+                       end
           i18n.data.merge! translated
           log_stderr t('i18n_tasks.translate_missing.translated', count: translated.leaves.count)
           print_forest translated, opt